PF-2301 - Improve footer's version info and add "data as of" in QA
This seems to work, but the futzing with the session and getting access to LastSQLServerDatabaseBackup wasn't as clean as I would like. It does work though, and should be unobtrusive. It failed a lot during development, so there is still a lot of failback code to deal with corner cases I may have missed, test environments, etc. In these cases you'll get a clearly bogus date ("1111-11-11").